@@ -41,17 +41,21 @@ Changelog = "https://github.com/beetbox/beets/blob/master/docs/changelog.rst"
 "Bug Tracker" = "https://github.com/beetbox/beets/issues"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<4"
+python = ">=3.10,<4"
 
 colorama = { version = "*", markers = "sys_platform == 'win32'" }
-confuse = ">=1.5.0"
+confuse = ">=2.1.0"
 jellyfish = "*"
 lap = ">=0.5.12"
 mediafile = ">=0.12.0"
-musicbrainzngs = ">=0.4"
-numpy = ">=1.24.4"
+numpy = [
+    { python = "<3.13", version = ">=2.0.2" },
+    { python = ">=3.13", version = ">=2.3.4" },
+]
 platformdirs = ">=3.5.0"
 pyyaml = "*"
+requests = ">=2.32.5"
+requests-ratelimiter = ">=0.7.0"
 typing_extensions = "*"
 unidecode = ">=1.3.6"
 
@@ -60,7 +64,16 @@ dbus-python = { version = "*", optional = true }
 flask = { version = "*", optional = true }
 flask-cors = { version = "*", optional = true }
 langdetect = { version = "*", optional = true }
-librosa = { version = "^0.10.2.post1", optional = true }
+librosa = { version = ">=0.11", optional = true }
+scipy = [ # for librosa
+    { python = "<3.13", version = ">=1.13.1", optional = true },
+    { python = ">=3.13", version = ">=1.16.1", optional = true },
+]
+musicbrainzngs = { version = ">=0.4", optional = true }
+numba = [ # for librosa
+    { python = "<3.13", version = ">=0.60", optional = true },
+    { python = ">=3.13", version = ">=0.62.1", optional = true },
+]
 mutagen = { version = ">=1.33", optional = true }
 Pillow = { version = "*", optional = true }
 py7zr = { version = "*", optional = true }
