@@ -16,24 +16,33 @@
 
 from __future__ import annotations
 
-import traceback
+import operator
 from collections import Counter
 from contextlib import suppress
-from functools import cached_property
-from itertools import product
-from typing import TYPE_CHECKING, Any, Iterable, Sequence
+from dataclasses import dataclass
+from functools import cached_property, singledispatchmethod
+from itertools import groupby, product
+from typing import TYPE_CHECKING, Any
 from urllib.parse import urljoin
 
-import musicbrainzngs
 from confuse.exceptions import NotFoundError
+from requests_ratelimiter import LimiterMixin
 
 import beets
 import beets.autotag.hooks
 from beets import config, plugins, util
 from beets.metadata_plugins import MetadataSourcePlugin
+from beets.util.deprecation import deprecate_for_user
 from beets.util.id_extractors import extract_release_id
 
+from ._utils.requests import (
+    HTTPNotFoundError,
+    RequestHandler,
+    TimeoutAndRetrySession,
+)
+
 if TYPE_CHECKING:
+    from collections.abc import Iterable, Sequence
     from typing import Literal
 
     from beets.library import Item
@@ -55,55 +64,33 @@ FIELDS_TO_MB_KEYS = {
     "year": "date",
 }
 
-musicbrainzngs.set_useragent("beets", beets.__version__, "https://beets.io/")
 
+RELEASE_INCLUDES = [
+    "artists",
+    "media",
+    "recordings",
+    "release-groups",
+    "labels",
+    "artist-credits",
+    "aliases",
+    "recording-level-rels",
+    "work-rels",
+    "work-level-rels",
+    "artist-rels",
+    "isrcs",
+    "url-rels",
+    "release-rels",
+    "genres",
+    "tags",
+]
 
-class MusicBrainzAPIError(util.HumanReadableError):
-    """An error while talking to MusicBrainz. The `query` field is the
-    parameter to the action and may have any type.
-    """
-
-    def __init__(self, reason, verb, query, tb=None):
-        self.query = query
-        if isinstance(reason, musicbrainzngs.WebServiceError):
-            reason = "MusicBrainz not reachable"
-        super().__init__(reason, verb, tb)
-
-    def get_message(self):
-        return f"{self._reasonstr()} in {self.verb} with query {self.query!r}"
-
-
-RELEASE_INCLUDES = list(
-    {
-        "artists",
-        "media",
-        "recordings",
-        "release-groups",
-        "labels",
-        "artist-credits",
-        "aliases",
-        "recording-level-rels",
-        "work-rels",
-        "work-level-rels",
-        "artist-rels",
-        "isrcs",
-        "url-rels",
-        "release-rels",
-        "tags",
-    }
-    & set(musicbrainzngs.VALID_INCLUDES["release"])
-)
-
-TRACK_INCLUDES = list(
-    {
-        "artists",
-        "aliases",
-        "isrcs",
-        "work-level-rels",
-        "artist-rels",
-    }
-    & set(musicbrainzngs.VALID_INCLUDES["recording"])
-)
+TRACK_INCLUDES = [
+    "artists",
+    "aliases",
+    "isrcs",
+    "work-level-rels",
+    "artist-rels",
+]
 
 BROWSE_INCLUDES = [
     "artist-credits",
@@ -112,19 +99,98 @@ BROWSE_INCLUDES = [
     "recording-rels",
     "release-rels",
 ]
-if "work-level-rels" in musicbrainzngs.VALID_BROWSE_INCLUDES["recording"]:
-    BROWSE_INCLUDES.append("work-level-rels")
 BROWSE_CHUNKSIZE = 100
 BROWSE_MAXTRACKS = 500
 
 
-def _preferred_alias(aliases: list[JSONDict]):
-    """Given an list of alias structures for an artist credit, select
-    and return the user's preferred alias alias or None if no matching
-    alias is found.
+class LimiterTimeoutSession(LimiterMixin, TimeoutAndRetrySession):
+    pass
+
+
+@dataclass
+class MusicBrainzAPI(RequestHandler):
+    api_host: str
+    rate_limit: float
+
+    def create_session(self) -> LimiterTimeoutSession:
+        return LimiterTimeoutSession(per_second=self.rate_limit)
+
+    def get_entity(
+        self, entity: str, inc_list: list[str] | None = None, **kwargs
+    ) -> JSONDict:
+        if inc_list:
+            kwargs["inc"] = "+".join(inc_list)
+
+        return self._group_relations(
+            self.get_json(
+                f"{self.api_host}/ws/2/{entity}",
+                params={**kwargs, "fmt": "json"},
+            )
+        )
+
+    def get_release(self, id_: str) -> JSONDict:
+        return self.get_entity(f"release/{id_}", inc_list=RELEASE_INCLUDES)
+
+    def get_recording(self, id_: str) -> JSONDict:
+        return self.get_entity(f"recording/{id_}", inc_list=TRACK_INCLUDES)
+
+    def browse_recordings(self, **kwargs) -> list[JSONDict]:
+        kwargs.setdefault("limit", BROWSE_CHUNKSIZE)
+        kwargs.setdefault("inc_list", BROWSE_INCLUDES)
+        return self.get_entity("recording", **kwargs)["recordings"]
+
+    @singledispatchmethod
+    @classmethod
+    def _group_relations(cls, data: Any) -> Any:
+        """Normalize MusicBrainz 'relations' into type-keyed fields recursively.
+
+        This helper rewrites payloads that use a generic 'relations' list into
+        a structure that is easier to consume downstream. When a mapping
+        contains 'relations', those entries are regrouped by their 'target-type'
+        and stored under keys like '<target-type>-relations'. The original
+        'relations' key is removed to avoid ambiguous access patterns.
+
+        The transformation is applied recursively so that nested objects and
+        sequences are normalized consistently, while non-container values are
+        left unchanged.
+        """
+        return data
+
+    @_group_relations.register(list)
+    @classmethod
+    def _(cls, data: list[Any]) -> list[Any]:
+        return [cls._group_relations(i) for i in data]
+
+    @_group_relations.register(dict)
+    @classmethod
+    def _(cls, data: JSONDict) -> JSONDict:
+        for k, v in list(data.items()):
+            if k == "relations":
+                get_target_type = operator.methodcaller("get", "target-type")
+                for target_type, group in groupby(
+                    sorted(v, key=get_target_type), get_target_type
+                ):
+                    relations = [
+                        {k: v for k, v in item.items() if k != "target-type"}
+                        for item in group
+                    ]
+                    data[f"{target_type}-relations"] = cls._group_relations(
+                        relations
+                    )
+                data.pop("relations")
+            else:
+                data[k] = cls._group_relations(v)
+        return data
+
+
+def _preferred_alias(
+    aliases: list[JSONDict], languages: list[str] | None = None
+) -> JSONDict | None:
+    """Given a list of alias structures for an artist credit, select
+    and return the user's preferred alias or None if no matching
     """
     if not aliases:
-        return
+        return None
 
     # Only consider aliases that have locales set.
     valid_aliases = [a for a in aliases if "locale" in a]
@@ -134,15 +200,18 @@ def _preferred_alias(aliases: list[JSONDict]):
     ignored_alias_types = [a.lower() for a in ignored_alias_types]
 
     # Search configured locales in order.
-    for locale in config["import"]["languages"].as_str_seq():
+    if languages is None:
+        languages = config["import"]["languages"].as_str_seq()
+
+    for locale in languages:
         # Find matching primary aliases for this locale that are not
         # being ignored
         matches = []
         for alias in valid_aliases:
             if (
                 alias["locale"] == locale
-                and "primary" in alias
-                and alias.get("type", "").lower() not in ignored_alias_types
+                and alias.get("primary")
+                and (alias.get("type") or "").lower() not in ignored_alias_types
             ):
                 matches.append(alias)
 
@@ -151,6 +220,8 @@ def _preferred_alias(aliases: list[JSONDict]):
             continue
 
         return matches[0]
+
+    return None
 
 
 def _multi_artist_credit(
@@ -164,36 +235,33 @@ def _multi_artist_credit(
     artist_sort_parts = []
     artist_credit_parts = []
     for el in credit:
-        if isinstance(el, str):
-            # Join phrase.
-            if include_join_phrase:
-                artist_parts.append(el)
-                artist_credit_parts.append(el)
-                artist_sort_parts.append(el)
+        alias = _preferred_alias(el["artist"].get("aliases", ()))
 
+        # An artist.
+        if alias:
+            cur_artist_name = alias["name"]
         else:
-            alias = _preferred_alias(el["artist"].get("alias-list", ()))
+            cur_artist_name = el["artist"]["name"]
+        artist_parts.append(cur_artist_name)
 
-            # An artist.
-            if alias:
-                cur_artist_name = alias["alias"]
-            else:
-                cur_artist_name = el["artist"]["name"]
-            artist_parts.append(cur_artist_name)
+        # Artist sort name.
+        if alias:
+            artist_sort_parts.append(alias["sort-name"])
+        elif "sort-name" in el["artist"]:
+            artist_sort_parts.append(el["artist"]["sort-name"])
+        else:
+            artist_sort_parts.append(cur_artist_name)
 
-            # Artist sort name.
-            if alias:
-                artist_sort_parts.append(alias["sort-name"])
-            elif "sort-name" in el["artist"]:
-                artist_sort_parts.append(el["artist"]["sort-name"])
-            else:
-                artist_sort_parts.append(cur_artist_name)
+        # Artist credit.
+        if "name" in el:
+            artist_credit_parts.append(el["name"])
+        else:
+            artist_credit_parts.append(cur_artist_name)
 
-            # Artist credit.
-            if "name" in el:
-                artist_credit_parts.append(el["name"])
-            else:
-                artist_credit_parts.append(cur_artist_name)
+        if include_join_phrase and (joinphrase := el.get("joinphrase")):
+            artist_parts.append(joinphrase)
+            artist_sort_parts.append(joinphrase)
+            artist_credit_parts.append(joinphrase)
 
     return (
         artist_parts,
@@ -263,9 +331,9 @@ def _preferred_release_event(
     ].as_str_seq()
 
     for country in preferred_countries:
-        for event in release.get("release-event-list", {}):
+        for event in release.get("release-events", {}):
             try:
-                if country in event["area"]["iso-3166-1-code-list"]:
+                if country in event["area"]["iso-3166-1-codes"]:
                     return country, event["date"]
             except KeyError:
                 pass
@@ -297,33 +365,9 @@ def _set_date_str(
                 setattr(info, key, date_num)
 
 
-def _is_translation(r):
-    _trans_key = "transl-tracklisting"
-    return r["type"] == _trans_key and r["direction"] == "backward"
-
-
-def _find_actual_release_from_pseudo_release(
-    pseudo_rel: JSONDict,
-) -> JSONDict | None:
-    try:
-        relations = pseudo_rel["release"]["release-relation-list"]
-    except KeyError:
-        return None
-
-    # currently we only support trans(liter)ation's
-    translations = [r for r in relations if _is_translation(r)]
-
-    if not translations:
-        return None
-
-    actual_id = translations[0]["target"]
-
-    return musicbrainzngs.get_release_by_id(actual_id, RELEASE_INCLUDES)
-
-
 def _merge_pseudo_and_actual_album(
     pseudo: beets.autotag.hooks.AlbumInfo, actual: beets.autotag.hooks.AlbumInfo
-) -> beets.autotag.hooks.AlbumInfo | None:
+) -> beets.autotag.hooks.AlbumInfo:
     """
     Merges a pseudo release with its actual release.
 
@@ -362,6 +406,24 @@ def _merge_pseudo_and_actual_album(
 
 
 class MusicBrainzPlugin(MetadataSourcePlugin):
+    @cached_property
+    def genres_field(self) -> str:
+        return f"{self.config['genres_tag'].as_choice(['genre', 'tag'])}s"
+
+    @cached_property
+    def api(self) -> MusicBrainzAPI:
+        hostname = self.config["host"].as_str()
+        if hostname == "musicbrainz.org":
+            hostname, rate_limit = "https://musicbrainz.org", 1.0
+        else:
+            https = self.config["https"].get(bool)
+            hostname = f"http{'s' if https else ''}://{hostname}"
+            rate_limit = (
+                self.config["ratelimit"].get(int)
+                / self.config["ratelimit_interval"].as_number()
+            )
+        return MusicBrainzAPI(hostname, rate_limit)
+
     def __init__(self):
         """Set up the python-musicbrainz-ngs module according to settings
         from the beets configuration. This should be called at startup.
@@ -374,6 +436,7 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
                 "ratelimit": 1,
                 "ratelimit_interval": 1,
                 "genres": False,
+                "genres_tag": "genre",
                 "external_ids": {
                     "discogs": False,
                     "bandcamp": False,
@@ -389,20 +452,11 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             self.config["search_limit"] = self.config["match"][
                 "searchlimit"
             ].get()
-            self._log.warning(
-                "'musicbrainz.searchlimit' option is deprecated and will be "
-                "removed in 3.0.0. Use 'musicbrainz.search_limit' instead."
+            deprecate_for_user(
+                self._log,
+                "'musicbrainz.searchlimit' configuration option",
+                "'musicbrainz.search_limit'",
             )
-        hostname = self.config["host"].as_str()
-        https = self.config["https"].get(bool)
-        # Only call set_hostname when a custom server is configured. Since
-        # musicbrainz-ngs connects to musicbrainz.org with HTTPS by default
-        if hostname != "musicbrainz.org":
-            musicbrainzngs.set_hostname(hostname, https)
-        musicbrainzngs.set_rate_limit(
-            self.config["ratelimit_interval"].as_number(),
-            self.config["ratelimit"].get(int),
-        )
 
     def track_info(
         self,
@@ -449,9 +503,9 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             info.artists_ids = _artist_ids(recording["artist-credit"])
             info.artist_id = info.artists_ids[0]
 
-        if recording.get("artist-relation-list"):
+        if recording.get("artist-relations"):
             info.remixer = _get_related_artist_names(
-                recording["artist-relation-list"], relation_type="remixer"
+                recording["artist-relations"], relation_type="remixer"
             )
 
         if recording.get("length"):
@@ -459,13 +513,13 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
 
         info.trackdisambig = recording.get("disambiguation")
 
-        if recording.get("isrc-list"):
-            info.isrc = ";".join(recording["isrc-list"])
+        if recording.get("isrcs"):
+            info.isrc = ";".join(recording["isrcs"])
 
         lyricist = []
         composer = []
         composer_sort = []
-        for work_relation in recording.get("work-relation-list", ()):
+        for work_relation in recording.get("work-relations", ()):
             if work_relation["type"] != "performance":
                 continue
             info.work = work_relation["work"]["title"]
@@ -474,7 +528,7 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
                 info.work_disambig = work_relation["work"]["disambiguation"]
 
             for artist_relation in work_relation["work"].get(
-                "artist-relation-list", ()
+                "artist-relations", ()
             ):
                 if "type" in artist_relation:
                     type = artist_relation["type"]
@@ -492,7 +546,7 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             info.composer_sort = ", ".join(composer_sort)
 
         arranger = []
-        for artist_relation in recording.get("artist-relation-list", ()):
+        for artist_relation in recording.get("artist-relations", ()):
             if "type" in artist_relation:
                 type = artist_relation["type"]
                 if type == "arranger":
@@ -524,9 +578,9 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             release["artist-credit"], include_join_phrase=False
         )
 
-        ntracks = sum(len(m["track-list"]) for m in release["medium-list"])
+        ntracks = sum(len(m["tracks"]) for m in release["media"])
 
-        # The MusicBrainz API omits 'artist-relation-list' and 'work-relation-list'
+        # The MusicBrainz API omits 'relations'
         # when the release has more than 500 tracks. So we use browse_recordings
         # on chunks of tracks to recover the same information in this case.
         if ntracks > BROWSE_MAXTRACKS:
@@ -535,35 +589,30 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             for i in range(0, ntracks, BROWSE_CHUNKSIZE):
                 self._log.debug("Retrieving tracks starting at {}", i)
                 recording_list.extend(
-                    musicbrainzngs.browse_recordings(
-                        release=release["id"],
-                        limit=BROWSE_CHUNKSIZE,
-                        includes=BROWSE_INCLUDES,
-                        offset=i,
-                    )["recording-list"]
+                    self.api.browse_recordings(release=release["id"], offset=i)
                 )
             track_map = {r["id"]: r for r in recording_list}
-            for medium in release["medium-list"]:
-                for recording in medium["track-list"]:
+            for medium in release["media"]:
+                for recording in medium["tracks"]:
                     recording_info = track_map[recording["recording"]["id"]]
                     recording["recording"] = recording_info
 
         # Basic info.
         track_infos = []
         index = 0
-        for medium in release["medium-list"]:
+        for medium in release["media"]:
             disctitle = medium.get("title")
             format = medium.get("format")
 
             if format in config["match"]["ignored_media"].as_str_seq():
                 continue
 
-            all_tracks = medium["track-list"]
+            all_tracks = medium["tracks"]
             if (
-                "data-track-list" in medium
+                "data-tracks" in medium
                 and not config["match"]["ignore_data_tracks"]
             ):
-                all_tracks += medium["data-track-list"]
+                all_tracks += medium["data-tracks"]
             track_count = len(all_tracks)
 
             if "pregap" in medium:
@@ -578,7 +627,7 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
 
                 if (
                     "video" in track["recording"]
-                    and track["recording"]["video"] == "true"
+                    and track["recording"]["video"]
                     and config["match"]["ignore_video_tracks"]
                 ):
                     continue
@@ -632,7 +681,7 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             artists=artists_names,
             artists_ids=album_artist_ids,
             tracks=track_infos,
-            mediums=len(release["medium-list"]),
+            mediums=len(release["media"]),
             artist_sort=artist_sort_name,
             artists_sort=artists_sort_names,
             artist_credit=artist_credit_name,
@@ -672,9 +721,9 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             rel_primarytype = release["release-group"]["primary-type"]
             if rel_primarytype:
                 albumtypes.append(rel_primarytype.lower())
-        if "secondary-type-list" in release["release-group"]:
-            if release["release-group"]["secondary-type-list"]:
-                for sec_type in release["release-group"]["secondary-type-list"]:
+        if "secondary-types" in release["release-group"]:
+            if release["release-group"]["secondary-types"]:
+                for sec_type in release["release-group"]["secondary-types"]:
                     albumtypes.append(sec_type.lower())
         info.albumtypes = albumtypes
 
@@ -690,8 +739,8 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
         _set_date_str(info, release_group_date, True)
 
         # Label name.
-        if release.get("label-info-list"):
-            label_info = release["label-info-list"][0]
+        if release.get("label-info"):
+            label_info = release["label-info"][0]
             if label_info.get("label"):
                 label = label_info["label"]["name"]
                 if label != "[no label]":
@@ -705,18 +754,18 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             info.language = rep.get("language")
 
         # Media (format).
-        if release["medium-list"]:
+        if release["media"]:
             # If all media are the same, use that medium name
-            if len({m.get("format") for m in release["medium-list"]}) == 1:
-                info.media = release["medium-list"][0].get("format")
+            if len({m.get("format") for m in release["media"]}) == 1:
+                info.media = release["media"][0].get("format")
             # Otherwise, let's just call it "Media"
             else:
                 info.media = "Media"
 
         if self.config["genres"]:
             sources = [
-                release["release-group"].get("tag-list", []),
-                release.get("tag-list", []),
+                release["release-group"].get(self.genres_field, []),
+                release.get(self.genres_field, []),
             ]
             genres: Counter[str] = Counter()
             for source in sources:
@@ -732,11 +781,11 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
         wanted_sources = {
             site for site, wanted in external_ids.items() if wanted
         }
-        if wanted_sources and (url_rels := release.get("url-relation-list")):
+        if wanted_sources and (url_rels := release.get("url-relations")):
             urls = {}
 
             for source, url in product(wanted_sources, url_rels):
-                if f"{source}.com" in (target := url["target"]):
+                if f"{source}.com" in (target := url["url"]["resource"]):
                     urls[source] = target
                     self._log.debug(
                         "Found link to {} release via MusicBrainz",
@@ -774,17 +823,20 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
     def get_album_criteria(
         self, items: Sequence[Item], artist: str, album: str, va_likely: bool
     ) -> dict[str, str]:
-        criteria = {
-            "release": album,
-            "alias": album,
-            "tracks": str(len(items)),
-        } | ({"arid": VARIOUS_ARTISTS_ID} if va_likely else {"artist": artist})
+        criteria = {"release": album} | (
+            {"arid": VARIOUS_ARTISTS_ID} if va_likely else {"artist": artist}
+        )
 
         for tag, mb_field in self.extra_mb_field_by_tag.items():
-            most_common, _ = util.plurality(i.get(tag) for i in items)
-            value = str(most_common)
-            if tag == "catalognum":
-                value = value.replace(" ", "")
+            if tag == "tracks":
+                value = str(len(items))
+            elif tag == "alias":
+                value = album
+            else:
+                most_common, _ = util.plurality(i.get(tag) for i in items)
+                value = str(most_common)
+                if tag == "catalognum":
+                    value = value.replace(" ", "")
 
             criteria[mb_field] = value
 
@@ -801,20 +853,17 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
         using the provided criteria. Handles API errors by converting them into
         MusicBrainzAPIError exceptions with contextual information.
         """
-        filters = {
-            k: _v for k, v in filters.items() if (_v := v.lower().strip())
-        }
-        self._log.debug(
-            "Searching for MusicBrainz {}s with: {!r}", query_type, filters
+        query = " AND ".join(
+            f'{k}:"{_v}"'
+            for k, v in filters.items()
+            if (_v := v.lower().strip())
         )
-        try:
-            method = getattr(musicbrainzngs, f"search_{query_type}s")
-            res = method(limit=self.config["search_limit"].get(), **filters)
-        except musicbrainzngs.MusicBrainzError as exc:
-            raise MusicBrainzAPIError(
-                exc, f"{query_type} search", filters, traceback.format_exc()
-            )
-        return res[f"{query_type}-list"]
+        self._log.debug(
+            "Searching for MusicBrainz {}s with: {!r}", query_type, query
+        )
+        return self.api.get_entity(
+            query_type, query=query, limit=self.config["search_limit"].get()
+        )[f"{query_type}s"]
 
     def candidates(
         self,
@@ -826,7 +875,10 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
         criteria = self.get_album_criteria(items, artist, album, va_likely)
         release_ids = (r["id"] for r in self._search_api("release", criteria))
 
-        yield from filter(None, map(self.album_for_id, release_ids))
+        for id_ in release_ids:
+            with suppress(HTTPNotFoundError):
+                if album_info := self.album_for_id(id_):
+                    yield album_info
 
     def item_candidates(
         self, item: Item, artist: str, title: str
@@ -849,29 +901,27 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             self._log.debug("Invalid MBID ({}).", album_id)
             return None
 
-        try:
-            res = musicbrainzngs.get_release_by_id(albumid, RELEASE_INCLUDES)
+        res = self.api.get_release(albumid)
 
-            # resolve linked release relations
-            actual_res = None
+        # resolve linked release relations
+        actual_res = None
 
-            if res["release"].get("status") == "Pseudo-Release":
-                actual_res = _find_actual_release_from_pseudo_release(res)
-
-        except musicbrainzngs.ResponseError:
-            self._log.debug("Album ID match failed.")
-            return None
-        except musicbrainzngs.MusicBrainzError as exc:
-            raise MusicBrainzAPIError(
-                exc, "get release by ID", albumid, traceback.format_exc()
-            )
+        if res.get("status") == "Pseudo-Release" and (
+            relations := res.get("release-relations")
+        ):
+            for rel in relations:
+                if (
+                    rel["type"] == "transl-tracklisting"
+                    and rel["direction"] == "backward"
+                ):
+                    actual_res = self.api.get_release(rel["release"]["id"])
 
         # release is potentially a pseudo release
-        release = self.album_info(res["release"])
+        release = self.album_info(res)
 
         # should be None unless we're dealing with a pseudo release
         if actual_res is not None:
-            actual_release = self.album_info(actual_res["release"])
+            actual_release = self.album_info(actual_res)
             return _merge_pseudo_and_actual_album(release, actual_release)
         else:
             return release
@@ -886,13 +936,7 @@ class MusicBrainzPlugin(MetadataSourcePlugin):
             self._log.debug("Invalid MBID ({}).", track_id)
             return None
 
-        try:
-            res = musicbrainzngs.get_recording_by_id(trackid, TRACK_INCLUDES)
-        except musicbrainzngs.ResponseError:
-            self._log.debug("Track ID match failed.")
-            return None
-        except musicbrainzngs.MusicBrainzError as exc:
-            raise MusicBrainzAPIError(
-                exc, "get recording by ID", trackid, traceback.format_exc()
-            )
-        return self.track_info(res["recording"])
+        with suppress(HTTPNotFoundError):
+            return self.track_info(self.api.get_recording(trackid))
+
+        return None
