@@ -64,8 +64,10 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                 }
             ],
             "date": "3001",
-            "medium-list": [],
-            "label-info-list": [
+            "media": [],
+            "genres": [{"count": 1, "name": "GENRE"}],
+            "tags": [{"count": 1, "name": "TAG"}],
+            "label-info": [
                 {
                     "catalog-number": "CATALOG NUMBER",
                     "label": {"name": "LABEL NAME"},
@@ -81,7 +83,7 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
         }
 
         if multi_artist_credit:
-            release["artist-credit"].append(" & ")  # add join phase
+            release["artist-credit"][0]["joinphrase"] = " & "
             release["artist-credit"].append(
                 {
                     "artist": {
@@ -122,7 +124,7 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                     ]
 
                     if multi_artist_credit:
-                        track["artist-credit"].append(" & ")  # add join phase
+                        track["artist-credit"][0]["joinphrase"] = " & "
                         track["artist-credit"].append(
                             {
                                 "artist": {
@@ -146,11 +148,11 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                     "number": "A1",
                 }
                 data_track_list.append(data_track)
-        release["medium-list"].append(
+        release["media"].append(
             {
                 "position": "1",
-                "track-list": track_list,
-                "data-track-list": data_track_list,
+                "tracks": track_list,
+                "data-tracks": data_track_list,
                 "format": medium_format,
                 "title": "MEDIUM TITLE",
             }
@@ -186,7 +188,7 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                 }
             ]
             if multi_artist_credit:
-                track["artist-credit"].append(" & ")  # add join phase
+                track["artist-credit"][0]["joinphrase"] = " & "
                 track["artist-credit"].append(
                     {
                         "artist": {
@@ -198,11 +200,10 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                     }
                 )
         if remixer:
-            track["artist-relation-list"] = [
+            track["artist-relations"] = [
                 {
                     "type": "remixer",
                     "type-id": "RELATION TYPE ID",
-                    "target": "RECORDING REMIXER ARTIST ID",
                     "direction": "RECORDING RELATION DIRECTION",
                     "artist": {
                         "id": "RECORDING REMIXER ARTIST ID",
@@ -213,7 +214,7 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                 }
             ]
         if video:
-            track["video"] = "true"
+            track["video"] = True
         if disambiguation:
             track["disambiguation"] = disambiguation
         return track
@@ -299,10 +300,10 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
                 "number": "A1",
             }
         ]
-        release["medium-list"].append(
+        release["media"].append(
             {
                 "position": "2",
-                "track-list": second_track_list,
+                "tracks": second_track_list,
             }
         )
 
@@ -515,6 +516,26 @@ class MBAlbumInfoTest(MusicBrainzTestCase):
         d = self.mb.album_info(release)
         assert d.data_source == "MusicBrainz"
 
+    def test_genres(self):
+        config["musicbrainz"]["genres"] = True
+        config["musicbrainz"]["genres_tag"] = "genre"
+        release = self._make_release()
+        d = self.mb.album_info(release)
+        assert d.genre == "GENRE"
+
+    def test_tags(self):
+        config["musicbrainz"]["genres"] = True
+        config["musicbrainz"]["genres_tag"] = "tag"
+        release = self._make_release()
+        d = self.mb.album_info(release)
+        assert d.genre == "TAG"
+
+    def test_no_genres(self):
+        config["musicbrainz"]["genres"] = False
+        release = self._make_release()
+        d = self.mb.album_info(release)
+        assert d.genre is None
+
     def test_ignored_media(self):
         config["match"]["ignored_media"] = ["IGNORED1", "IGNORED2"]
         tracks = [
@@ -678,15 +699,15 @@ class ArtistFlatteningTest(unittest.TestCase):
 
     def _add_alias(self, credit_dict, suffix="", locale="", primary=False):
         alias = {
-            "alias": f"ALIAS{suffix}",
+            "name": f"ALIAS{suffix}",
             "locale": locale,
             "sort-name": f"ALIASSORT{suffix}",
         }
         if primary:
             alias["primary"] = "primary"
-        if "alias-list" not in credit_dict["artist"]:
-            credit_dict["artist"]["alias-list"] = []
-        credit_dict["artist"]["alias-list"].append(alias)
+        if "aliases" not in credit_dict["artist"]:
+            credit_dict["artist"]["aliases"] = []
+        credit_dict["artist"]["aliases"].append(alias)
 
     def test_single_artist(self):
         credit = [self._credit_dict()]
@@ -703,7 +724,10 @@ class ArtistFlatteningTest(unittest.TestCase):
         assert c == ["CREDIT"]
 
     def test_two_artists(self):
-        credit = [self._credit_dict("a"), " AND ", self._credit_dict("b")]
+        credit = [
+            {**self._credit_dict("a"), "joinphrase": " AND "},
+            self._credit_dict("b"),
+        ]
         a, s, c = musicbrainz._flatten_artist_credit(credit)
         assert a == "NAMEa AND NAMEb"
         assert s == "SORTa AND SORTb"
@@ -761,86 +785,86 @@ class MBLibraryTest(MusicBrainzTestCase):
     def test_follow_pseudo_releases(self):
         side_effect = [
             {
-                "release": {
-                    "title": "pseudo",
-                    "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
-                    "status": "Pseudo-Release",
-                    "medium-list": [
-                        {
-                            "track-list": [
-                                {
-                                    "id": "baz",
-                                    "recording": {
-                                        "title": "translated title",
-                                        "id": "bar",
-                                        "length": 42,
-                                    },
-                                    "position": 9,
-                                    "number": "A1",
-                                }
-                            ],
-                            "position": 5,
-                        }
-                    ],
-                    "artist-credit": [
-                        {
-                            "artist": {
-                                "name": "some-artist",
-                                "id": "some-id",
-                            },
-                        }
-                    ],
-                    "release-group": {
-                        "id": "another-id",
-                    },
-                    "release-relation-list": [
-                        {
-                            "type": "transl-tracklisting",
-                            "target": "d2a6f856-b553-40a0-ac54-a321e8e2da01",
-                            "direction": "backward",
-                        }
-                    ],
-                }
+                "title": "pseudo",
+                "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
+                "status": "Pseudo-Release",
+                "media": [
+                    {
+                        "tracks": [
+                            {
+                                "id": "baz",
+                                "recording": {
+                                    "title": "translated title",
+                                    "id": "bar",
+                                    "length": 42,
+                                },
+                                "position": 9,
+                                "number": "A1",
+                            }
+                        ],
+                        "position": 5,
+                    }
+                ],
+                "artist-credit": [
+                    {
+                        "artist": {
+                            "name": "some-artist",
+                            "id": "some-id",
+                        },
+                    }
+                ],
+                "release-group": {
+                    "id": "another-id",
+                },
+                "release-relations": [
+                    {
+                        "type": "transl-tracklisting",
+                        "direction": "backward",
+                        "release": {
+                            "id": "d2a6f856-b553-40a0-ac54-a321e8e2da01"
+                        },
+                    }
+                ],
             },
             {
-                "release": {
-                    "title": "actual",
-                    "id": "d2a6f856-b553-40a0-ac54-a321e8e2da01",
-                    "status": "Official",
-                    "medium-list": [
-                        {
-                            "track-list": [
-                                {
-                                    "id": "baz",
-                                    "recording": {
-                                        "title": "original title",
-                                        "id": "bar",
-                                        "length": 42,
-                                    },
-                                    "position": 9,
-                                    "number": "A1",
-                                }
-                            ],
-                            "position": 5,
-                        }
-                    ],
-                    "artist-credit": [
-                        {
-                            "artist": {
-                                "name": "some-artist",
-                                "id": "some-id",
-                            },
-                        }
-                    ],
-                    "release-group": {
-                        "id": "another-id",
-                    },
-                    "country": "COUNTRY",
-                }
+                "title": "actual",
+                "id": "d2a6f856-b553-40a0-ac54-a321e8e2da01",
+                "status": "Official",
+                "media": [
+                    {
+                        "tracks": [
+                            {
+                                "id": "baz",
+                                "recording": {
+                                    "title": "original title",
+                                    "id": "bar",
+                                    "length": 42,
+                                },
+                                "position": 9,
+                                "number": "A1",
+                            }
+                        ],
+                        "position": 5,
+                    }
+                ],
+                "artist-credit": [
+                    {
+                        "artist": {
+                            "name": "some-artist",
+                            "id": "some-id",
+                        },
+                    }
+                ],
+                "release-group": {
+                    "id": "another-id",
+                },
+                "country": "COUNTRY",
             },
         ]
 
-        with mock.patch("musicbrainzngs.get_release_by_id") as gp:
+        with mock.patch(
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_release"
+        ) as gp:
             gp.side_effect = side_effect
             album = self.mb.album_for_id("d2a6f856-b553-40a0-ac54-a321e8e2da02")
             assert album.country == "COUNTRY"
@@ -848,44 +872,43 @@ class MBLibraryTest(MusicBrainzTestCase):
     def test_pseudo_releases_with_empty_links(self):
         side_effect = [
             {
-                "release": {
-                    "title": "pseudo",
-                    "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
-                    "status": "Pseudo-Release",
-                    "medium-list": [
-                        {
-                            "track-list": [
-                                {
-                                    "id": "baz",
-                                    "recording": {
-                                        "title": "translated title",
-                                        "id": "bar",
-                                        "length": 42,
-                                    },
-                                    "position": 9,
-                                    "number": "A1",
-                                }
-                            ],
-                            "position": 5,
-                        }
-                    ],
-                    "artist-credit": [
-                        {
-                            "artist": {
-                                "name": "some-artist",
-                                "id": "some-id",
-                            },
-                        }
-                    ],
-                    "release-group": {
-                        "id": "another-id",
-                    },
-                    "release-relation-list": [],
-                }
-            },
+                "title": "pseudo",
+                "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
+                "status": "Pseudo-Release",
+                "media": [
+                    {
+                        "tracks": [
+                            {
+                                "id": "baz",
+                                "recording": {
+                                    "title": "translated title",
+                                    "id": "bar",
+                                    "length": 42,
+                                },
+                                "position": 9,
+                                "number": "A1",
+                            }
+                        ],
+                        "position": 5,
+                    }
+                ],
+                "artist-credit": [
+                    {
+                        "artist": {
+                            "name": "some-artist",
+                            "id": "some-id",
+                        },
+                    }
+                ],
+                "release-group": {
+                    "id": "another-id",
+                },
+            }
         ]
 
-        with mock.patch("musicbrainzngs.get_release_by_id") as gp:
+        with mock.patch(
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_release"
+        ) as gp:
             gp.side_effect = side_effect
             album = self.mb.album_for_id("d2a6f856-b553-40a0-ac54-a321e8e2da02")
             assert album.country is None
@@ -893,43 +916,43 @@ class MBLibraryTest(MusicBrainzTestCase):
     def test_pseudo_releases_without_links(self):
         side_effect = [
             {
-                "release": {
-                    "title": "pseudo",
-                    "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
-                    "status": "Pseudo-Release",
-                    "medium-list": [
-                        {
-                            "track-list": [
-                                {
-                                    "id": "baz",
-                                    "recording": {
-                                        "title": "translated title",
-                                        "id": "bar",
-                                        "length": 42,
-                                    },
-                                    "position": 9,
-                                    "number": "A1",
-                                }
-                            ],
-                            "position": 5,
-                        }
-                    ],
-                    "artist-credit": [
-                        {
-                            "artist": {
-                                "name": "some-artist",
-                                "id": "some-id",
-                            },
-                        }
-                    ],
-                    "release-group": {
-                        "id": "another-id",
-                    },
-                }
-            },
+                "title": "pseudo",
+                "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
+                "status": "Pseudo-Release",
+                "media": [
+                    {
+                        "tracks": [
+                            {
+                                "id": "baz",
+                                "recording": {
+                                    "title": "translated title",
+                                    "id": "bar",
+                                    "length": 42,
+                                },
+                                "position": 9,
+                                "number": "A1",
+                            }
+                        ],
+                        "position": 5,
+                    }
+                ],
+                "artist-credit": [
+                    {
+                        "artist": {
+                            "name": "some-artist",
+                            "id": "some-id",
+                        },
+                    }
+                ],
+                "release-group": {
+                    "id": "another-id",
+                },
+            }
         ]
 
-        with mock.patch("musicbrainzngs.get_release_by_id") as gp:
+        with mock.patch(
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_release"
+        ) as gp:
             gp.side_effect = side_effect
             album = self.mb.album_for_id("d2a6f856-b553-40a0-ac54-a321e8e2da02")
             assert album.country is None
@@ -937,50 +960,52 @@ class MBLibraryTest(MusicBrainzTestCase):
     def test_pseudo_releases_with_unsupported_links(self):
         side_effect = [
             {
-                "release": {
-                    "title": "pseudo",
-                    "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
-                    "status": "Pseudo-Release",
-                    "medium-list": [
-                        {
-                            "track-list": [
-                                {
-                                    "id": "baz",
-                                    "recording": {
-                                        "title": "translated title",
-                                        "id": "bar",
-                                        "length": 42,
-                                    },
-                                    "position": 9,
-                                    "number": "A1",
-                                }
-                            ],
-                            "position": 5,
-                        }
-                    ],
-                    "artist-credit": [
-                        {
-                            "artist": {
-                                "name": "some-artist",
-                                "id": "some-id",
-                            },
-                        }
-                    ],
-                    "release-group": {
-                        "id": "another-id",
-                    },
-                    "release-relation-list": [
-                        {
-                            "type": "remaster",
-                            "target": "d2a6f856-b553-40a0-ac54-a321e8e2da01",
-                            "direction": "backward",
-                        }
-                    ],
-                }
-            },
+                "title": "pseudo",
+                "id": "d2a6f856-b553-40a0-ac54-a321e8e2da02",
+                "status": "Pseudo-Release",
+                "media": [
+                    {
+                        "tracks": [
+                            {
+                                "id": "baz",
+                                "recording": {
+                                    "title": "translated title",
+                                    "id": "bar",
+                                    "length": 42,
+                                },
+                                "position": 9,
+                                "number": "A1",
+                            }
+                        ],
+                        "position": 5,
+                    }
+                ],
+                "artist-credit": [
+                    {
+                        "artist": {
+                            "name": "some-artist",
+                            "id": "some-id",
+                        },
+                    }
+                ],
+                "release-group": {
+                    "id": "another-id",
+                },
+                "release-relations": [
+                    {
+                        "type": "remaster",
+                        "direction": "backward",
+                        "release": {
+                            "id": "d2a6f856-b553-40a0-ac54-a321e8e2da01"
+                        },
+                    }
+                ],
+            }
         ]
 
-        with mock.patch("musicbrainzngs.get_release_by_id") as gp:
+        with mock.patch(
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_release"
+        ) as gp:
             gp.side_effect = side_effect
             album = self.mb.album_for_id("d2a6f856-b553-40a0-ac54-a321e8e2da02")
             assert album.country is None
@@ -1025,15 +1050,13 @@ class TestMusicBrainzPlugin(PluginMixin):
 
         assert mb.get_album_criteria(items, "Artist ", " Album", va_likely) == {
             "release": " Album",
-            "alias": " Album",
-            "tracks": str(len(items)),
             **expected_additional_criteria,
         }
 
     def test_item_candidates(self, monkeypatch, mb):
         monkeypatch.setattr(
-            "musicbrainzngs.search_recordings",
-            lambda *_, **__: {"recording-list": [self.RECORDING]},
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_json",
+            lambda *_, **__: {"recordings": [self.RECORDING]},
         )
 
         candidates = list(mb.item_candidates(Item(), "hello", "there"))
@@ -1043,34 +1066,32 @@ class TestMusicBrainzPlugin(PluginMixin):
 
     def test_candidates(self, monkeypatch, mb):
         monkeypatch.setattr(
-            "musicbrainzngs.search_releases",
-            lambda *_, **__: {"release-list": [{"id": self.mbid}]},
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_json",
+            lambda *_, **__: {"releases": [{"id": self.mbid}]},
         )
         monkeypatch.setattr(
-            "musicbrainzngs.get_release_by_id",
+            "beetsplug.musicbrainz.MusicBrainzAPI.get_release",
             lambda *_, **__: {
-                "release": {
-                    "title": "hi",
-                    "id": self.mbid,
-                    "status": "status",
-                    "medium-list": [
-                        {
-                            "track-list": [
-                                {
-                                    "id": "baz",
-                                    "recording": self.RECORDING,
-                                    "position": 9,
-                                    "number": "A1",
-                                }
-                            ],
-                            "position": 5,
-                        }
-                    ],
-                    "artist-credit": [
-                        {"artist": {"name": "some-artist", "id": "some-id"}}
-                    ],
-                    "release-group": {"id": "another-id"},
-                }
+                "title": "hi",
+                "id": self.mbid,
+                "status": "status",
+                "media": [
+                    {
+                        "tracks": [
+                            {
+                                "id": "baz",
+                                "recording": self.RECORDING,
+                                "position": 9,
+                                "number": "A1",
+                            }
+                        ],
+                        "position": 5,
+                    }
+                ],
+                "artist-credit": [
+                    {"artist": {"name": "some-artist", "id": "some-id"}}
+                ],
+                "release-group": {"id": "another-id"},
             },
         )
         candidates = list(mb.candidates([], "hello", "there", False))
@@ -1078,3 +1099,84 @@ class TestMusicBrainzPlugin(PluginMixin):
         assert len(candidates) == 1
         assert candidates[0].tracks[0].track_id == self.RECORDING["id"]
         assert candidates[0].album == "hi"
+
+
+def test_group_relations():
+    raw_release = {
+        "id": "r1",
+        "relations": [
+            {"target-type": "artist", "type": "vocal", "name": "A"},
+            {"target-type": "url", "type": "streaming", "url": "http://s"},
+            {"target-type": "url", "type": "purchase", "url": "http://p"},
+            {
+                "target-type": "work",
+                "type": "performance",
+                "work": {
+                    "relations": [
+                        {
+                            "artist": {"name": "幾田りら"},
+                            "target-type": "artist",
+                            "type": "composer",
+                        },
+                        {
+                            "target-type": "url",
+                            "type": "lyrics",
+                            "url": {
+                                "resource": "https://utaten.com/lyric/tt24121002/"
+                            },
+                        },
+                        {
+                            "artist": {"name": "幾田りら"},
+                            "target-type": "artist",
+                            "type": "lyricist",
+                        },
+                        {
+                            "target-type": "url",
+                            "type": "lyrics",
+                            "url": {
+                                "resource": "https://www.uta-net.com/song/366579/"
+                            },
+                        },
+                    ],
+                    "title": "百花繚乱",
+                    "type": "Song",
+                },
+            },
+        ],
+    }
+
+    assert musicbrainz.MusicBrainzAPI._group_relations(raw_release) == {
+        "id": "r1",
+        "artist-relations": [{"type": "vocal", "name": "A"}],
+        "url-relations": [
+            {"type": "streaming", "url": "http://s"},
+            {"type": "purchase", "url": "http://p"},
+        ],
+        "work-relations": [
+            {
+                "type": "performance",
+                "work": {
+                    "artist-relations": [
+                        {"type": "composer", "artist": {"name": "幾田りら"}},
+                        {"type": "lyricist", "artist": {"name": "幾田りら"}},
+                    ],
+                    "url-relations": [
+                        {
+                            "type": "lyrics",
+                            "url": {
+                                "resource": "https://utaten.com/lyric/tt24121002/"
+                            },
+                        },
+                        {
+                            "type": "lyrics",
+                            "url": {
+                                "resource": "https://www.uta-net.com/song/366579/"
+                            },
+                        },
+                    ],
+                    "title": "百花繚乱",
+                    "type": "Song",
+                },
+            },
+        ],
+    }
